reformatted datetime string

+++ lib/libvol2bird.c
@@ -3261,7 +3261,7 @@ void writeCSV(char *filename, vol2bird_t* alldata, char* source, char* fileIn, c
         nanify(profileBio[11+iCopied]), nanify(profileBio[12+iCopied]), // eta, dens
         nanify(profileBio[9+iCopied]),nanify(profileAll[9+iCopied]), // dbz, dbz_all
         (int)nanify(profileBio[10+iCopied]),(int)nanify(profileBio[13+iCopied]), // n, n_dbz 
-        nanify(profileAll[10+iCopied]),nanify(profileAll[13+iCopied]), // n_all, n_dbz_all
+        (int)nanify(profileAll[10+iCopied]),(int)nanify(profileAll[13+iCopied]), // n_all, n_dbz_all
         *rcs, *sd_vvp_thresh, *vcp);
         // rcs , sd_vvp_threshold
         // vcp, radar_latitude, radar_longitude, radar_height, radar_wavelength, source_file
